# Faster node endpoint