--- a/config/urls.py
+++ b/config/urls.py
@@ -13,6 +13,7 @@ Including another URLconf
     1. Import the include() function: from django.urls import include, path
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
+
 from django.conf import settings
 from django.contrib import admin
 from django.urls import path, include
@@ -39,5 +40,9 @@ urlpatterns = [
         name="oauth2-redirect",
     ),
     path("downloads/", include("downloads.urls")),
-    path("_debug/", include("debug_toolbar.urls"))
 ]
+
+if settings.DEBUG:
+    urlpatterns += [
+        path("_debug/", include("debug_toolbar.urls")),
+    ]