--- a/requirements/base.txt
+++ b/requirements/base.txt
@@ -16,3 +16,6 @@ django-filter==23.5
 git+https://github.com/waggle-sensor/django-address
 # setuptools needed for python 3.12+
 setuptools==69.2.0
+# NOTE(sean) I'm installing django-debug-toolbar in base to make the collectstatic step work
+# when building the image. We can revisit this later if we want to move it to just prod.
+django-debug-toolbar==4.3.0 # enables a debug toolbar, helps with debugging api endpoints